requirements.txt
numpy>=1.17
qiskit-terra>=0.17.0
pandas
dataclasses
